api/requirements.txt
flask
openai
chromadb
keybert
spacy
soundfile
llama-cpp-python